port r19347 from trunk: fix warnings

+++ src/core/disk_ops/disk_unistd_o_direct.c
@@ -82,7 +82,7 @@ static void *starpu_unistd_o_direct_plug(void *parameter, starpu_ssize_t size)
 void *starpu_unistd_o_direct_global_async_read(void *base, void *obj, void *buf, off_t offset, size_t size)
 {
 	STARPU_ASSERT_MSG((size % getpagesize()) == 0, "The unistd_o_direct variant can only read a multiple of page size %lu Bytes (Here %lu). Use the non-o_direct unistd variant if your data is not a multiple of %lu",
-			  getpagesize(), (unsigned long) size, getpagesize());
+			(unsigned long) getpagesize(), (unsigned long) size, (unsigned long) getpagesize());
 
 	STARPU_ASSERT_MSG((((uintptr_t) buf) % getpagesize()) == 0, "You have to use starpu_malloc function");
 
@@ -92,7 +92,7 @@ void *starpu_unistd_o_direct_global_async_read(void *base, void *obj, void *buf,
 void *starpu_unistd_o_direct_global_async_write(void *base, void *obj, void *buf, off_t offset, size_t size)
 {
 	STARPU_ASSERT_MSG((size % getpagesize()) == 0, "The unistd_o_direct variant can only write a multiple of page size %lu Bytes (Here %lu). Use the non-o_direct unistd variant if your data is not a multiple of %lu",
-			  getpagesize(), (unsigned long) size, getpagesize());
+			(unsigned long) getpagesize(), (unsigned long) size, (unsigned long) getpagesize());
 
 	STARPU_ASSERT_MSG((((uintptr_t)buf) % getpagesize()) == 0, "You have to use starpu_malloc function");
